style: rename apis
