+++ src/ts/type/index.ts
@@ -4,4 +4,4 @@ export * from './Jwt'
 export * from './User'
 export * from './Enums'
 export * from './Snackbar'
-export * from './Api'
+export * from './Apis'
